use long for paginator results and plurals (#6786)

+++ app/templating/StringHelper.scala
@@ -11,7 +11,7 @@ trait StringHelper { self: NumberHelper =>
 
   def shorten(text: String, length: Int, sep: String = "…") = lila.common.String.shorten(text, length, sep)
 
-  def pluralize(s: String, n: Int) = s"$n $s${if (n > 1) "s" else ""}"
+  def pluralize(s: String, n: Long) = s"$n $s${if (n > 1) "s" else ""}"
 
   def showNumber(n: Int): String = if (n > 0) s"+$n" else n.toString
