+++ modules/i18n/src/main/package.scala
@@ -4,7 +4,7 @@ import play.api.i18n.Lang
 
 package object i18n extends PackageObject {
 
-  type Count      = Int
+  type Count      = Long
   type MessageKey = String
 
   private[i18n] type MessageMap = java.util.Map[MessageKey, Translation]
