+++ modules/search/src/main/SearchReadApi.scala
@@ -4,5 +4,5 @@ trait SearchReadApi[A, Q] {
 
   def search(query: Q, from: From, size: Size): Fu[List[A]]
 
-  def count(query: Q): Fu[Int]
+  def count(query: Q): Fu[Long]
 }
